fix: use absolute path for volumes
One issue encountered has docker failing to run the tempest container.
This is due to the command assuming that the current working directory
is the root of kayobe-config which is not true if you run
./deploy-openstack.sh from the home directory.

Also provide sometime before attempting to attach to the docker
container via the SSH command. Finally, don't prompt to accept host keys
as this expected to run automatically.