feat: simplify commands using `ansible_user`
Some commands such as accessing the control host via SSH or running the
playbooks relies on the `ansible_user` to be defined which is typically
either `ubuntu` or `cloud-user`.

To make this easy we can reuse and export the variable `ssh_user` which
is assigned by the user within `terraform.tfvars`. This should be set to
the user for accessing the machine. Once set and `terraform apply`
executed users can ssh using the command on line 215 verbatim. The same
can be said for the ansible-playbook commands.