Fix issue building kayobe image
A regression was introduced in:

https://github.com/stackhpc/terraform-kayobe-multinode/pull/55

which meant that the Kayobe docker image failed to build. This is
because requirements.txt was updated to point a local path that did not
exist inside the docker image during the build. I've changed the logic
to update the requirements.txt to use the ansible variables that
configure the kayobe checkout.