Work around Ansible rc-13 race condition (mk2)
There is a race condition in Ansible that can result in this failure:
  msg: |-
  MODULE FAILURE
  See stdout/stderr for the exact error
rc: -13
See https://github.com/ansible/ansible/issues/78344 and
https://github.com/ansible/ansible/issues/81777.

In https://github.com/stackhpc/stackhpc-kayobe-config/pull/1108 we applied
a workaround to increase the ControlPersist timeout to 1 hour, but this
does not always work.

Here we use a different workaround of disabling SSH pipelining. This has
performance implications for Ansible, but is a reasonable trade-off for
reliability.

We set the config option as an environment variable rather than in
ansible.cfg in Kayobe configuration, to avoid a merge conflict on upgrade.