Workaround: Disable SSH pipelining
There is a race condition in Ansible that can result in this failure:
  msg: |-
  MODULE FAILURE
  See stdout/stderr for the exact error
rc: -13

See https://github.com/ansible/ansible/issues/78344 and
https://github.com/ansible/ansible/issues/81777.

In https://github.com/stackhpc/stackhpc-kayobe-config/pull/1108 we
applied a workaround to increase the ControlPersist timeout to 1 hour,
but this does not always work.

Disabling SSH pipelining prevents the issue at the cost of Ansible
execution duration.