Make adding internal TLS config to globals.yml idempotent
Previously the internal TLS config in globals-tls-config.yml was always
appended to Kolla Ansible's globals.yml. This could result in duplicated
variables.

This change only appends the config when it is not already present in
globals.yml.

This also makes it possible to add the internal TLS config permanently
within globals.yml, behind a conditional.